adjust the way the sum of signals is calculated
reuse value of previously calculated sum of signals
unless the list of IonMeasurements is changed

+++ openchrom/plugins/net.openchrom.msd.converter.supplier.cms/src/net/openchrom/msd/converter/supplier/cms/model/ICalibratedVendorMassSpectrum.java
@@ -36,7 +36,7 @@ public interface ICalibratedVendorMassSpectrum extends ICalibratedVendorLibraryM
 
 	String getScanName();
 
-	double getSumSignal();
+	float getSumSignal();
 
 	ICalibratedVendorMassSpectrum makeDeepCopy() throws CloneNotSupportedException;
